move exporter logic to dedicated package
This allows external projects (like github.com/grafana/agent) to embed
the exporter.
