+++ main_test.go
@@ -17,7 +17,6 @@ import (
 	"bytes"
 	"context"
 	"io/ioutil"
-	"net"
 	"net/http"
 	"os"
 	"os/exec"
@@ -26,9 +25,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/go-kit/kit/log"
 	"github.com/grobie/gomemcache/memcache"
-	"github.com/prometheus/client_golang/prometheus"
 )
 
 func TestAcceptance(t *testing.T) {
@@ -204,56 +201,4 @@ OUTER:
 	cancel()
 
 	<-errc
-}
-
-func TestParseStatsSettings(t *testing.T) {
-	addr, err := net.ResolveIPAddr("ip4", "127.0.0.1")
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	t.Run("Success", func(t *testing.T) {
-		t.Parallel()
-
-		var statsSettings = map[net.Addr]map[string]string{
-			addr: {
-				"maxconns":              "10",
-				"lru_crawler":           "yes",
-				"lru_crawler_sleep":     "100",
-				"lru_crawler_tocrawl":   "0",
-				"lru_maintainer_thread": "no",
-				"hot_lru_pct":           "20",
-				"warm_lru_pct":          "40",
-				"hot_max_factor":        "0.20",
-				"warm_max_factor":       "2.00",
-			},
-		}
-		ch := make(chan prometheus.Metric, 100)
-		e := NewExporter("", 100*time.Millisecond, log.NewNopLogger())
-		if err := e.parseStatsSettings(ch, statsSettings); err != nil {
-			t.Errorf("expect return error, error: %v", err)
-		}
-	})
-
-	t.Run("Failure", func(t *testing.T) {
-		t.Parallel()
-		var statsSettings = map[net.Addr]map[string]string{
-			addr: {
-				"maxconns":              "10",
-				"lru_crawler":           "yes",
-				"lru_crawler_sleep":     "100",
-				"lru_crawler_tocrawl":   "0",
-				"lru_maintainer_thread": "fail",
-				"hot_lru_pct":           "20",
-				"warm_lru_pct":          "40",
-				"hot_max_factor":        "0.20",
-				"warm_max_factor":       "2.00",
-			},
-		}
-		ch := make(chan prometheus.Metric, 100)
-		e := NewExporter("", 100*time.Millisecond, log.NewNopLogger())
-		if err := e.parseStatsSettings(ch, statsSettings); err == nil {
-			t.Error("expect return error but not")
-		}
-	})
 }
